TO for filtering options fixed

+++ commons/src/main/java/com/dzydowicz/meteoritelandings/tos/MeteoriteLandingFilterTO.java
@@ -15,6 +15,6 @@ public class MeteoriteLandingFilterTO {
     private MeteoriteLandingFallEnum fall;
     private Float minMass;
     private Float maxMass;
-    private Float minYear;
-    private Float maxYear;
+    private Integer minYear;
+    private Integer maxYear;
 }
